requirements.txt
pytest
selenium==2.53.6
suds-jurko